Resolves: tdf#103174 & rhbz#1367846 improve gtk3 trackpad scrolling
convert number of "lines" scrolled to double and allow
fractional parts of lines/columns


Related: rhbz#1367846 queue and merge scroll events

Reviewed-on: https://gerrit.libreoffice.org/37779
Tested-by: Caolán McNamara <caolanm@redhat.com>
(cherry picked from commit 7f60978b2ccd0e17816b78bde60c6e0e60a9d52e)

Reviewed-on: https://gerrit.libreoffice.org/37913
Tested-by: Jenkins <ci@libreoffice.org>
(cherry picked from commit f7d2bf216afa10268e6a7c1d4613a2fd8f7c7f3c)

+++ include/vcl/window.hxx
@@ -719,7 +719,7 @@ private:
     SAL_DLLPRIVATE void                 ImplCallActivateListeners(vcl::Window*);
     SAL_DLLPRIVATE void                 ImplCallDeactivateListeners(vcl::Window*);
 
-    SAL_DLLPRIVATE static void          ImplHandleScroll( ScrollBar* pHScrl, long nX, ScrollBar* pVScrl, long nY );
+    SAL_DLLPRIVATE static void          ImplHandleScroll(ScrollBar* pHScrl, double nX, ScrollBar* pVScrl, double nY);
 
     SAL_DLLPRIVATE Rectangle            ImplOutputToUnmirroredAbsoluteScreenPixel( const Rectangle& rRect ) const;
     SAL_DLLPRIVATE long                 ImplGetUnmirroredOutOffX();
